+++ src/app/index.js
@@ -1,20 +1,13 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { createStore, applyMiddleware } from 'redux'
 import { Provider } from 'react-redux'
-import thunk from 'redux-thunk'
-import reducer from './reducers'
 
+import store from './store'
 import App from './components/App'
 
 import 'normalize.css'
 import 'codemirror/lib/codemirror.css'
 import './styles.css'
-
-const store = createStore(
-  reducer,
-  applyMiddleware(thunk)
-)
 
 ReactDOM.render(
   <Provider store={store}>