+++ src/app/components/Output/index.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import Codemirror from 'react-codemirror'
+import CopyButton from '../CopyButton'
 import 'codemirror/mode/javascript/javascript'
 
 import './styles.css'
@@ -15,6 +16,7 @@ export default function ({ children, theme }) {
   return (
     <div className='Output'>
       <Codemirror options={options} value={children} />
+      <CopyButton value={children} />
     </div>
   )
 }