+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Set the version of CoffeeScript to target
 - Set the theme for the editor
 - Set the theme for the output
+- Settings are persisted between app restarts
 
 # 1.0.0